Move ConsumerAllocationTests to correct folder
