+++ x/ccv/consumer/keeper/distribution_test.go
@@ -12,7 +12,6 @@ import (
 
 	testkeeper "github.com/cosmos/interchain-security/v4/testutil/keeper"
 	"github.com/cosmos/interchain-security/v4/x/ccv/consumer/types"
-	providertypes "github.com/cosmos/interchain-security/v4/x/ccv/provider/types"
 	ccvtypes "github.com/cosmos/interchain-security/v4/x/ccv/types"
 )
 
@@ -90,40 +89,4 @@ func TestAllowedRewardDenoms(t *testing.T) {
 	require.Len(t, allowedDenoms, 2)
 	require.Equal(t, allowedDenoms[0], "ustake")
 	require.True(t, strings.HasPrefix(allowedDenoms[1], "ibc/"))
-}
-
-func TestSetConsumerRewardsAllocation(t *testing.T) {
-	keeperParams := testkeeper.NewInMemKeeperParams(t)
-	ctx := keeperParams.Ctx
-
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	mocks := testkeeper.NewMockedKeepers(ctrl)
-	providerKeeper := testkeeper.NewInMemProviderKeeper(keeperParams, mocks)
-
-	rewardAllocation := providertypes.ConsumerRewardsAllocation{
-		Rewards: sdk.NewDecCoins(sdk.NewDecCoin("uatom", sdk.NewInt(1000))),
-	}
-
-	providerKeeper.SetConsumerRewardsAllocation(ctx, "consumer-1", rewardAllocation)
-
-	alloc := providerKeeper.GetConsumerRewardsAllocation(ctx, "consumer-1")
-	require.Equal(t, rewardAllocation, alloc)
-}
-
-func TestGetConsumerRewardsAllocationNil(t *testing.T) {
-	keeperParams := testkeeper.NewInMemKeeperParams(t)
-	ctx := keeperParams.Ctx
-
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	mocks := testkeeper.NewMockedKeepers(ctrl)
-	providerKeeper := testkeeper.NewInMemProviderKeeper(keeperParams, mocks)
-
-	alloc := providerKeeper.GetConsumerRewardsAllocation(ctx, "consumer-1")
-
-	expectedRewardAllocation := providertypes.ConsumerRewardsAllocation{
-		Rewards: nil,
-	}
-	require.Equal(t, expectedRewardAllocation, alloc)
 }
