+++ x/ccv/provider/keeper/distribution_test.go
@@ -16,6 +16,7 @@ import (
 
 	testkeeper "github.com/cosmos/interchain-security/v4/testutil/keeper"
 	"github.com/cosmos/interchain-security/v4/x/ccv/provider/types"
+	providertypes "github.com/cosmos/interchain-security/v4/x/ccv/provider/types"
 )
 
 func TestComputeConsumerTotalVotingPower(t *testing.T) {
@@ -232,4 +233,40 @@ func TestIdentifyConsumerChainIDFromIBCPacket(t *testing.T) {
 			}
 		})
 	}
+}
+
+func TestSetConsumerRewardsAllocation(t *testing.T) {
+	keeperParams := testkeeper.NewInMemKeeperParams(t)
+	ctx := keeperParams.Ctx
+
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+	mocks := testkeeper.NewMockedKeepers(ctrl)
+	providerKeeper := testkeeper.NewInMemProviderKeeper(keeperParams, mocks)
+
+	rewardAllocation := providertypes.ConsumerRewardsAllocation{
+		Rewards: sdk.NewDecCoins(sdk.NewDecCoin("uatom", sdk.NewInt(1000))),
+	}
+
+	providerKeeper.SetConsumerRewardsAllocation(ctx, "consumer-1", rewardAllocation)
+
+	alloc := providerKeeper.GetConsumerRewardsAllocation(ctx, "consumer-1")
+	require.Equal(t, rewardAllocation, alloc)
+}
+
+func TestGetConsumerRewardsAllocationNil(t *testing.T) {
+	keeperParams := testkeeper.NewInMemKeeperParams(t)
+	ctx := keeperParams.Ctx
+
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+	mocks := testkeeper.NewMockedKeepers(ctrl)
+	providerKeeper := testkeeper.NewInMemProviderKeeper(keeperParams, mocks)
+
+	alloc := providerKeeper.GetConsumerRewardsAllocation(ctx, "consumer-1")
+
+	expectedRewardAllocation := providertypes.ConsumerRewardsAllocation{
+		Rewards: nil,
+	}
+	require.Equal(t, expectedRewardAllocation, alloc)
 }
